refactor: update lint, add lint docs, update git hooks
