+++ .gitignore
@@ -1,3 +1,5 @@
+ktlint
+
 # Built application files
 *.apk
 *.aar
