+++ lefthook.yml
@@ -1,0 +1,18 @@
+# Configuration for lefthook tool to manage git hooks 
+# https://github.com/evilmartians/lefthook
+
+
+# Commands to run before committing code. 
+pre-commit:
+  commands:    
+    lint:
+      glob: "*.kt"
+      # Format kotlin code and then `git add` modified files from formatter. 
+      run: make format && git add {staged_files} 
+
+# Commands to run before pushing code 
+pre-push:
+  commands:
+    lint: 
+      # Run linter giving you errors to fix. 
+      run: make lint-no-format
