+++ messagingpush/src/sharedTest/java/io/customer/messagingpush/ModuleMessagingPushFCMTest.kt
@@ -1,7 +1,7 @@
 package io.customer.messagingpush
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import io.customer.common_test.BaseTest
+import io.customer.commontest.BaseTest
 import io.customer.messagingpush.provider.FCMTokenProvider
 import io.customer.sdk.CustomerIOInstance
 import io.customer.sdk.utils.random
