+++ sdk/src/main/java/io/customer/sdk/extensions/EnumExtensions.kt
@@ -1,3 +1,4 @@
+/* ktlint-disable filename */ // until this extension file contains 2+ functions in it, we will disable this ktlint rule.
 package io.customer.sdk.extensions
 
 /**
