+++ sdk/src/sharedTest/java/io/customer/sdk/queue/QueueStorageIntegrationTest.kt
@@ -2,14 +2,14 @@ package io.customer.sdk.queue
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.customer.base.extenstions.subtract
+import io.customer.commontest.BaseTest
 import io.customer.sdk.queue.type.QueueInventory
 import io.customer.sdk.queue.type.QueueModifyResult
 import io.customer.sdk.queue.type.QueueStatus
+import io.customer.sdk.queue.type.QueueTaskGroup
 import io.customer.sdk.queue.type.QueueTaskMetadata
 import io.customer.sdk.queue.type.QueueTaskRunResults
 import io.customer.sdk.utils.random
-import io.customer.common_test.BaseTest
-import io.customer.sdk.queue.type.QueueTaskGroup
 import org.amshove.kluent.*
 import org.junit.Before
 import org.junit.Test
